Reconnect on errors, including send errors
In general, it's not possible to recover after gen_tcp:send
has returned an error, so we reconnect in this case.

When receiving tcp_error and ssl_error instead of waiting
for a tcp_closed. An error is not always followed by a close
e.g. if the network is down.

Ignore messages and errors for an old/wrong socket.