Handle setopts results instead of crashing during connect
If setting socket options fails during a connection procedure
we now return the error instead of causing a crash.
This was the last known crash-point when connecting to a client.

If setting the socket active again after receiving some data
we now close the socket and reconnect (..if reconnection not disabled)

Starting an eredis_client with config `{reconnect_sleep, no_reconnect}`
still emits an exit signal when the connection fails.