Patch documentation to enable publishing on hexdocs.pm
Eredis docs are generated as markdown files using edown, but there seems
to be some formatting error when publishing documentation using Elixirs
mix and ex_doc.
To get readable documentation on hexdocs.pm we need to patch the markdown files,
and use the latest version of ex_doc. This will give similar result as the docs
on Github.

Added a build target to publish the package to hex.pm aswell.
ex_doc seems to need an empty file `doc/.build` which prohibits removal of
the markdown files before attempting to convert them to html..