optimize first clause of `eredis_parser:split_by_newlines`
Optimize the first clause for the most common case - when Acc is empty.
This happens a lot with small DB commands which do not produce much
output at all (which is most DB commands in almost every use case).

The new clause partially duplicates the 3rd clause,
but avoids building a new binary from the empty Acc.

Use guard `size() == 0` as it is the faster than matching on <<>>